Move implementation of packetbeat system tests to new base
Not all code was moved to far, as there are some special implementations in packetbeat
