+++ packetbeat/tests/system/test_0001_mysql_spaces.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
     def test_mysql_with_spaces(self):
         self.render_config_template(
             mysql_ports=[3306]
