+++ packetbeat/tests/system/test_0003_http_simple.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
 
     def test_http_sample(self):
         self.render_config_template()
