+++ packetbeat/tests/system/test_0004_ipv6.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
 
     def test_ipv6_thrift_framed(self):
         self.render_config_template(
