+++ packetbeat/tests/system/test_0009_pgsql.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
     def test_select(self):
         self.render_config_template(
             pgsql_ports=[5432],
