+++ packetbeat/tests/system/test_0010_http_10_connection_close.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
 
     def test_http_sample(self):
         self.render_config_template(http_ports=['8000'])
