+++ packetbeat/tests/system/test_0020_mysql_send_options.py
@@ -1,4 +1,4 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 """
 Tests for the mysql's send_request and send_response
@@ -6,7 +6,7 @@ options.
 """
 
 
-class Test(TestCase):
+class Test(BaseTest):
 
     def test_default_settings(self):
         """
