+++ packetbeat/tests/system/test_0025_mongodb_basic.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
     """
     Basic MongoDB tests
     """
