+++ packetbeat/tests/system/test_0027_mysql_affected_rows.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
 
     def test_mysql_affected_rows(self):
         self.render_config_template(
