+++ packetbeat/tests/system/test_0028_mysql_error.py
@@ -1,7 +1,7 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
-class Test(TestCase):
+class Test(BaseTest):
 
     def test_mysql_error(self):
         self.render_config_template(
