+++ packetbeat/tests/system/test_0041_memcache_udp_bin_basic.py
@@ -1,4 +1,4 @@
-from pbtests.packetbeat import TestCase
+from packetbeat import BaseTest
 
 
 import pprint
@@ -10,7 +10,7 @@ def pretty(*k, **kw):
     pp.pprint(*k, **kw)
 
 
-class Test(TestCase):
+class Test(BaseTest):
     def _run(self, pcap):
         self.render_config_template(
             memcache_udp_transaction_timeout=10
